Collapse app/input/ folder; move files into systems/ and util/ (refs #1200)
Per the ECS folder audit (#1200), app/input/ was a system-in-costume —
its .cpp files are dispatcher wiring + listener implementations that
belong with the rest of the systems, and its lone util header belongs
in app/util/.

Mechanical relocation, no behavior change:

  app/input/input_dispatcher.cpp     -> app/systems/input_dispatcher.cpp
  app/input/input_routing.h          -> app/systems/input_routing.h
  app/input/game_state_routing.cpp   -> app/systems/game_state_routing.cpp
  app/input/keyboard_shape_mapping.h -> app/util/keyboard_shape_mapping.h

The two relocated .cpp files are now picked up by the existing
SYSTEM_SOURCES glob, so the dedicated INPUT_SOURCES glob and lib
argument are dropped from CMakeLists.txt. The keyboard mapping is
header-only and needs no CMake change. app/input/ is removed.

Include sites updated in:
  app/game_loop.cpp, app/systems/input_system.cpp,
  tests/test_helpers.h, tests/test_input_pipeline_behavior.cpp,
  benchmarks/bench_systems.cpp
plus a stale-path comment in app/systems/game_state_system.cpp.

Validation: native macOS build succeeds with -Wall -Wextra -Werror;
all 901 Catch2 tests pass (913 ctest cases incl. node tests).
