+++ app/systems/input_system.cpp
@@ -1,7 +1,7 @@
 #include "all_systems.h"
 #include "camera_system.h"
 #include "web_input_policy.h"
-#include "../input/keyboard_shape_mapping.h"
+#include "../util/keyboard_shape_mapping.h"
 #include "../components/input.h"
 #include "../components/input_events.h"
 #include "../components/player.h"
